fixing controllers and refactoring some comparator folders. will finish everything later today
