+++ src/main/java/Entities/ProductDateComparator.java
@@ -1,5 +1,7 @@
 package Entities;
 
+import UseCases.Comparator.ProductComparator;
+
 /**
  * a class that compares 2 Products with respect to their added date
  */
